requirements.txt
# Azure OpenAI
openai>=1.0.0
langchain-openai>=0.1.0

# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
pytesseract==0.3.10

# Machine Learning
numpy==1.24.3
scikit-learn==1.3.2

# Voice processing
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11

# Web scraping and external search
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2